Current: Some simple UX, validation of fields, and error messages.
Next: Cleanup of data for better validation, Firebase accounts, more UX and comments.

Comment: I'm going to bed earlier today, to see if I can fix my schedule, and to take a bit of a break, maybe watch some anime, idk. https://youtu.be/2nWmBJxW9q0 (the video might be a bit cringe, but the song is majestic) I love when rock has an orchestra and goes theatrical. Hi from the past, and have a nice day! Take care and DRINK WATER!