Current: Limited chars of textfields, organized and cleaned up
Next: Complexity of password, full name for the users?, "Forgot Password" UI, Firebase accounts, more UX and comments.

Comment: I wasn't going to work on this today, I was going to "take a break", buuuuut... here I am ^.^
I just feel like I'm not doing shit if I'm just watching series all day, idk, and this felt productive xD. AS LONG AS I CAN EVADE AN EXISTENTIAL CRISIS then it's all good.
Hello from the past, hope you have/you're having/you had a great day! And here... https://youtu.be/HLiCZbhZdsM, an adventure awaits, but drink water and take care! << might be a bit cringe xD but it's because I'm listening to that song right now, so put it on 2:40 and read from "Hello from... take care!" and imagine me screaming and waving with a smile from half a mile a way as the sun sets. (I know, I'm very self-aware of how weird all this is xD, it is 5:05am, and I haven't slept, the life of a programmer? Maybe just the life of someone with a fucked up sleeping schedule) BYE!