Current: Some UX, Complexity of password, "Forgot Password" UI
Next: Firebase accounts, and comments.

Comment: Hello from the past, today I didn't do a lot because I felt tired and took a nap of 2 hours in the middle of the day .-.
But the next part should be interesting, creating accounts and setting the DB, to then start with the Quiz page 👌.
Here https://youtu.be/iu_0kOfMGD0, the change in 0:13-0:18 is godly!